[#51] feat: 돌아가기 버튼 공용 컴포넌트로 분리하여 구현
